New method of calculating weighted scattering weights to publish.
OVERVIEW:
    The majority of the discrepancy between AMFs calculated from a
weighted sum of clear and cloudy AMFs and one calculated from a
weighted sum of clear and cloudy scattering weights was due to the
combination of setting scattering weights below terrain or cloud
pressure to 0 and the interpolation.
    When calculating the combined scattering weights, the
interpolation to terrain or cloud pressure was done after setting the
scattering weights below that to 0; therefore the interpolation was
between f(i+1) and 0 rather than f(i+1) and f(i). This is corrected
for by returning vectors of scattering weights with additional
pressure levels at the terrain and cloud pressures. Everything below
those can then be set to 0 before making the combined scattering
weights.

MODIFIED FILES:
    integPr2 - now returns a pressure and mixing ratio/scattering
weight vector with interpolated values (these values are determined by
the fourth input, which is a matrix of pressures to interpolate to).
This ensures that the same interpolation process is used to reach
those points. Also, the interpolation process was moved into a nested
function.

    omiAmfAK2 - changed the computation of the "publishable"
scattering weights to use the new scattering weight vectors output
from integPr2. There is also a lot of testing code to check that the
AMFs produced from different scattering weights are the same; this
will be removed soon.

    BEHR_main - debugging changes, made to get it to run with the
current omiAmfAK2; these will be undone as soon as omiAmfAK2 is
finalized.