Changes to publish scattering weights, and update to hdf_quad template
MODIFIED FILES:
    integPr2 - now only gives a warning about NaNs not being included
in the column if there are at least some non-NaN values.

    omiAmfAK2 - now output the ghost correction factor as well.

    BEHR_main - restored parfor loop (instead of for). Added in the
ghost fraction and scattering weights to the Data structure, removed
the fields that used the secondary AMF for testing.

    BEHR_hdf_v2 - added in the BEHRGhostFraction variable.

    hdf_quadrangle_BEHR - removed the testing variable
BEHRColumnAmountNO2Trop_AMF2.

    hdf_quadreangle_template - made to take in an OMI structure (to
keep the field order the same) and make unedited grid cells NaNs as
the hdf_quadrangle_BEHR does.