Updates to prep to publish DISCOVER-AQ BEHR data.
RENAMED FILES:
    HDF Tools/BEHR_hdf_v2.m --> HDF Tools/BEHR_publishing_v2.m - renamed to
better reflect that it can now handle both HDF and text (csv) publishing.

NEW FILES:
    Read_Data/read_myd14.m - New function that can parse MODIS thermal anomaly
(MYD14) level 2 data. Because of the format of MYD14 products, it returns
several vectors specifying the latitude, longitude, FRP, and certainty of the
fire. (Not really related to publishing, but needed to be added.)

    Documentation/BEHR InSitu Readme/BEHR_InSitu_Readme.{tex,pdf} - a LaTeX
readme on the in situ reprocessed product, which will be posted to the website
for end users.

MODIFIED FILES:
    AMF_tools/omiAmfAK2.m - changed the ghost correction so that the AMF is
divided by the ghost multiplicative factor, rather than multiplied by it.
Justification:

        The ghost factor, G is given by G = V_gnd / (w * V_cld + (1-w) * V_gnd)
        where V_gnd and V_cld are the modeled VCD above the surface and cloud
        top respectively, and w is the cloud fraction.  Therefore, this quantity
        represents the ratio of the total column (visible + ghost) to the
        visible column only. Therefore, V_tot = V_vis * G, where V_vis is the
        observed visible column from OMI, and V_tot is naturally the total VCD.

        Let A_vis be the initial AMF calculated based on the cloudy conditions.
        Then V_vis = S / A_vis. Since V_tot = V_vis * G = (S / A_vis) * G,
        which can be rewritten as V_tot = S / (A_vis / G).

        Note that BEHR_main has also been changed to only report the visible VCD

    BEHR_Main/BEHR_main.m - now passed noGhost = 0 to omiAmfAK2, which will
cause it to only return the visible column. It will still return the ghost
factor for users who wish to add in the ghost column.

    Documentation/BEHR User Guide/BEHR-User-Guide.{tex,pdf} - added Section 3,
which enumerates and describes the quantities in the HDF files.

    Documentation/Changelog.txt - noted that the BEHRColumnAmountNO2Trop no
longer contains a ghost column.