General update of Utils folder
NEW FILES:
    Constants/omno2d_centers.m and Constants/omno2d_corners.m - return center or
corner points for the OMNO2d dataset (gridded OMI NO2).

    mat2latex.m - converts a matrix to the format necessary for Latex tables.

MODIFIED FILES:
    Constants/geos_chem_centers.m - allow selection of different resolutions.
2x2.5 and 0.3125x0.250 for N.Am. implemented.

    Constants/merge_field_names.m - updated some of the input checking and
modified the aerosol fields to use the blue extinction sent by Lee Thornhill.
Also added several biomass burning tracers to relevant campaigns, and a number
of range files for SEAC4RS. Finally, made it so that it will not ask for a
decision on which range file to use if a range file is not requested.

    plot_error_envelope_y.m - updates to the input parser, checking on the
validity of the optional inputs.  All required inputs converted to column
vectors so that Matlab treats each point in turn, rather than acting on a whole
row if a matrix is passed.

    quicksave.m - leaves spaces but removes forward slashes in names, other
fixes to keep the figure's knowledge of its filename intact.

    savefigs.m - include date of save in file name and replaces % with the word
"Percent"

    state_outlines.m - updates to handle figure handles in R2014b and later.