Work towards figuring out why the pseudo retrieval has an all-over +ve change in AMF
MODIFIED FILES:
    AMF_tools/omiAmfAK2.m - made the padding of the profile output arrays and
such able to handle any number of dimensions. I think this change also exists in
another branch already.

    Utils/cat_sat_data.m - added ability to concatenate 3D variable in the along
track direction, as well as the option to just concatenate along a new
dimension, which will only work if the arrays are always the same size.

    Utils/misc_behr_wind_plots.m - new function to plot the ensemble of a priori
profiles and adding the hybrid-avg case (the direct average of the hybrid a
priori) to the diff plotting functions.