Spate of updates that haven't been committed.
MODIFIED FILES:
    ask_multichoice.m - quick fix dealing with case sensitivity.

    ask_number.m - allows multiple numbers to be input. Also handles cases where
the user written test function is expecting a scalar value and didn't get one.
It still errors, but tries to be a little more informative.

    coord_mat.m - can create a matrix of any number of dimensions.

    find_nearest_gridpoint.m - help text updated.

    iscellcontents.m - can now accept either a string defining the test function
or an actual test function, for greater flexibility.

    mat2latex.m - successfully got it to allow you to force it to print numbers
without scientific notation when using the 'u' special format.

    quicksave.m - switched to use uiputfile which makes it much more convenient
for saving files with a similar name to another.

    savefigs.m - allows you to tell it to save all figures in one go, without
having to hit enter repeatedly.  Will have to use default file names in this
case.

    tex_in_printf.m - will now correctly identify special characters at the end
of a string.

NEW FILES:
    draw_arrow.m - allows you to draw an arrow on the current plot.

    padcat.m - concatenates different size matrices by padding them with NaNs by
default until the sizes match.

    randdist.m - allows random number sampling from a distribution defined by an
arbitrary probability distribution function.

    spxlim.m, spylim.m - reset the x- and y- limits respectively of all sub
plots in a figure.

    wrap_string.m - takes a string and wraps it to an 80 character wide
character array. The width can be changed.  It will try to break at spaces if
possible.