Adjusting how GLOBETerpres is handled, especially over oceans.
MODIFIED FILES:
    Read_Data/read_omno2_v_aug2012.m - previously when reading the GLOBE terrain
database, NaNs were replaced with -500 (as a fill value). However, NaNs are only
present over ocean, so it makes sense to just make them 0 (as ocean should be at
sea level). Also make it import GLOBE data for ~10 degrees around the lat/lon
limit to be sure we have terrain data where we have NO2 profiles. This causes
some small changes to coastal VCDs (due to averaging).

    BEHR_Main/BEHR_main.m - previously NaNs in both GLOBETerpres and MODISAlbedo
were replaced with 1013 and 0 respectively. MODISAlbedo only seems to be NaN
when the longitude for that pixel is, so there's no sense in giving it a value.
With GLOBETerpres, this made pixels outside of the domain appear to have a valid
surface pressure when in fact they do not, so this has been removed. This also
caused VCDs in Canada to change (pixels that had previously had surfP = 1013
were changed, fix included expanding the GLOBE domain in read_omno2).

    Utils/Constants/BEHR_version.m - version string updated to 2-1B.