Typical mass update of general Utils folder.
MODIFIED FILES:
    savefigs.m - help text added, also the capability to append a string to
every file name if using 'savefigs('all').

    state_outlines.m - can now just return state outlines as NaN-separated
vectors for use in manual plotting or other uses.

NEW FILES:
    Colormaps/*.m (blue_red_cmap, blue_red_only_cmap, blue_red_white_cmap,
four_color_cmap, valin_cmap). Five functions that return a color map array
(64x3). Can be used with the colormap function as e.g. colormap(blue_red_cmap)
to set the colormap of the current figure. Several are duplicated as .mat files
in the BEHR repo.

    cprintf.m - accepts a cell array of strings and will insert the additional
arguments into each string; shortcut of "for i=1:numel(C); C2{i} =
sprintf(C{i}, ...); end"

    sprintfmulti.m - counterpart to cprintf, takes a single string as the first
argument but can take cell arrays as the additional arguments, and will generate
a cell array in which each element of the input cell arrays is substituded into
the format string in turn.

    draw_circle.m - draws a circle on a plot.

    histsemilog.m - plots a histogram with a log scale on the y-axis. This works
around a bug wherein setting the y-scale of a plot generated by HIST() to 'log'
will cause the bars to all but vanish.

    mat2cell_simple.m - actually deprecated, only in here in case of
dependencies. Its functionality is actually handled by the built-in NUM2CELL,
using this function will issue that warning. Will be removed in the future.

    ncatts.m - function that will print out the name and value of attributes for
a variable (or the root) of a netCDF structure returned from ncinfo.

    xylims.m - function used to set x- and y- limits to be the same.