Updated multiple utilities
MODIFIED FILES:
    combine_plots.m - now copies legends

    find_submatrix2.m - allows the matrices to have NaNs as long as the start
and end points of the child matrix are not NaNs. Also fixed a bug with output
not being set if two output variables requested.

    label_axis_with_letter.m - documented better, and will automatically convert
an index 1-26 to the appropriate letter, assuming a format of (a). Can still
pass a string as the first argument to customize formatting. Removed the
'parent' parameter, since its function was already fullfilled by 'ax'

    mat2latex.m - can now output the string as a variable, rather than forcing
it to print to the screen. If no output requested, will print to screen.

    reldiff.m - option to compute the relative difference using the average of
the two values in the denominator added.

    tex_in_printf.m - extra blank line removed

NEW FILES:
    label_subfigs.m - wrapper around label_axis_with_letter that finds the child
axes and labels them.

    make_latex_table.m - makes a full table environment using mat2latex to
generate the table itself. It can handle row and column names, though not
multirow or multicolumn labels. It can also reture the whole resulting
environment as a string variable, save it to a new file, or insert it into an
existing file with appropriate marks in it.