Some clean up of the code, starting to work with SOAS aircraft data
NEW FILES:
    read_icart_files/combine_merges.m - combine two or more Merge structures;
can be used to either combine merges that represent different times in the same
day, have different data fields for the same times, or a combination or the two.

    read_icart_files/read_icartt_file.m - file that took the main ICARTT reading
code from read_merge_data.m so that it could be reused in other functions.

    utility_scripts/sitenum_from_profnum.m - helper function that takes a
profile number (which represents both which site the profile occured at and what
number the profile was within the campaign) and returns the site number.

    UnitTests/merge_unit_tests.m - unit tests for the "combine_merge" function
to ensure it behaves as expected.

MODIFIED FILES:
    read_icart_files/read_merge_data.m - the bulk of the code moved to the new
read_icartt_file() function, to facilitate its use in other code. This function
retains the code for selecting the ICARTT file(s) and saving the resulting .mat
files.

    read_icart_files/read_merge_field.m - old eval() calls to access arbitrary
field names replaced with the Struct.(fieldname) syntax. Proper handling of
number of output variables added. Auxilary values (UTC, altitude, longitude,
latitude) have NaNs inserted for data points that are fill values or similar in
the main data field.

    satellite_verification_scripts/Run_Spiral_Verification.m - help text updated
with the syntax for calls to this function. Can now take campaign name as an
input and handles the request for all profiles to be included (since profile
filtering was added to work with aerosol effects (needed to be able to only
collect profiles that met aerosol above/below/coincident with the NO2 layer)

    satellite_verification_scripts/spiral_verification_avg_pix2prof.m - when
reading albedo (solely for output in the debugging structure), favor the BRDF
albedo, if present.

    utility_scripts/campaign_wide_ops.m - some help text and input cleanup,
mainly added the ability to output a day-by-day average profile. Some of the
other binning function combined to eliminate redundant code.

    utility_scripts/merge_field_names.m - TexAQS and SOAS directories and dates
added.